Further split up workers and move to their own folder
