+++ rana_qgis_plugin/workers/persistent.py
