+++ rana_qgis_plugin/workers/styling.py
